[mac] Fixing tons of compiler warnings created by the mac compiler. All of them somewhat valid, heh?

+++ The-Orm/PatchButtonPanel.cpp
@@ -143,7 +143,7 @@ void PatchButtonPanel::setupPageButtons() {
 	}
 	// Any more buttons unused?
 	for (int i = 0; i < maxPageButtons_; i++) {
-		if (i >= pageButtonMap_.size()) {
+		if (static_cast<size_t>(i) >= pageButtonMap_.size()) {
 			pageNumbers_[i]->setVisible(false);
 		}
 	}
@@ -216,7 +216,7 @@ void PatchButtonPanel::refresh(bool async, int autoSelectTarget /* = -1 */) {
 
 	// Now set the button text and colors
 	int active = indexOfActive();
-	for (int i = 0; i < (int) std::max(patchButtons_->size(), patches_.size()); i++) {
+	for (size_t i = 0; i < std::max(patchButtons_->size(), patches_.size()); i++) {
 		if (i < patchButtons_->size()) {
 			auto button = patchButtons_->buttonWithIndex(i);
 			if (i < patches_.size() && patches_[i].patch() && patches_[i].synth()) {
